The 'filesize' is the most important one to update. The others are unlikely
to ever get close to the 4G limit (they count connections, or messages), but
it doesn't cost us much (and all the serialized forms are JSON strings, so
the wire form is identical anyways, as long as we don't go beyond `2**53`
where JSON cannot represent them accurately).

closes #69

+++ src/core/server_messages.rs
@@ -22,7 +22,7 @@ pub enum OutboundMessage {
     Open { mailbox: String },
     Add { phase: String, body: String },
     Close { mailbox: String, mood: String },
-    Ping { ping: u32 },
+    Ping { ping: u64 },
 }
 
 pub fn bind(appid: &AppID, side: &MySide) -> OutboundMessage {
@@ -78,7 +78,7 @@ pub fn deserialize_outbound(s: &str) -> OutboundMessage {
 }
 
 #[allow(dead_code)]
-pub fn ping(ping: u32) -> OutboundMessage {
+pub fn ping(ping: u64) -> OutboundMessage {
     OutboundMessage::Ping { ping }
 }
 
@@ -109,7 +109,7 @@ pub enum InboundMessage {
     Closed {},
     Ack {},
     Pong {
-        pong: u32,
+        pong: u64,
     },
     //Error { error: String, orig: Message },
 }
